+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * guess Parquet driver from `.parquet` file extension; #2506
 
-* move dependency `classInt` to Suggests:
+* adapt tests to changes in GEOS >= 3.15.0
 
 # version 1.1-2